# [SPARK-14267] [SQL] [PYSPARK] execute multiple Python UDFs within single batch