@davies 2016-03-31
@JoshRosen I had addressed you comments also pushed some changes to improve the performance.