@davies 2016-03-31
Merging this into master, thanks!